pyproject.toml
[project]
name = "tavily-search"
version = "0.1.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "httpx>=0.28.0",
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
 "tavily-python>=0.5.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
tavily-search = "tavily_search:main"
